requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dataclasses-json==0.6.7
environs==11.0.0
filelock==3.15.3
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
-e file:///home/robinb/dev/langchain/libs/langchain#egg=langchain
-e file:///home/robinb/dev/langchain/libs/community#egg=langchain_community
-e file:///home/robinb/dev/langchain/libs/core#egg=langchain_core
langchain-huggingface==0.0.3
langchain-milvus==0.1.1
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langsmith==0.1.81
MarkupSafe==2.1.5
marshmallow==3.21.3
milvus-lite==2.4.7
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.5.2.13
nvidia-cuda-cupti-cu12==12.5.39
nvidia-cuda-nvrtc-cu12==12.5.40
nvidia-cuda-runtime-cu12==12.5.39
nvidia-cudnn-cu12==9.1.1.17
nvidia-cufft-cu12==11.2.3.18
nvidia-curand-cu12==10.3.6.39
nvidia-cusolver-cu12==11.6.2.40
nvidia-cusparse-cu12==12.4.1.24
nvidia-nccl-cu12==2.22.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.5.39
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pip-review==1.3.0
protobuf==5.27.1
psutil==6.0.0
pydantic==2.7.4
pydantic_core==2.18.4
pymilvus==2.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-triton==3.0.0+45fff310c8
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==3.0.1
six==1.16.0
SQLAlchemy==2.0.31
sympy==1.12.1
tenacity==8.4.1
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.5.0.dev20240620+cu124
torchaudio==2.4.0.dev20240620+cu124
torchvision==0.20.0.dev20240620+cu124
tqdm==4.66.4
transformers==4.41.2
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
yarl==1.9.4
